Upper bounds on rates of double reduction

Calculates the upper bounds of the double reduction parameters according to the complete equation segregation model. See Huang et. al. (2019) for details.

drbounds(ploidy)

Arguments

ploidy
The ploidy of the species. Should be even and at least 4.

Value

A vector of length floor(ploidy/4). Element i is the upper bound on the probability of i pairs of identical-by-double-reduction alleles being in an individual.

References

Huang, K., Wang, T., Dunn, D. W., Zhang, P., Cao, X., Liu, R., & Li, B. (2019). Genotypic frequencies at equilibrium for polysomic inheritance under double-reduction. G3: Genes, Genomes, Genetics, 9(5), 1693-1706. doi: 10.1534/g3.119.400132

Author

David Gerard

Examples

drbounds(4)

#> [1] 0.1666667

drbounds(6)

#> [1] 0.3

drbounds(8)

#> [1] 0.38571429 0.02142857

drbounds(10)

#> [1] 0.43650794 0.05952381

drbounds(12)

#> [1] 0.462662338 0.105519481 0.002705628

drbounds(14)

#> [1] 0.47115385 0.15297203 0.01019814

drbounds(16)

#> [1] 0.4669386169 0.1978438228 0.0231157731 0.0003399378